requirements.txt
chess>=1.9.0
pandas>=1.3.0
scikit-learn>=1.0.0
openai>=1.0.0
numpy>=1.21.0
wandb>=0.21.0
hydra-core>=1.3.0
stockfish>=3.28.0
langchain-core>=0.1.0
langchain>=0.1.0
langchain_anthropic
langchain_openai
torchvision
torch
accelerate
transformers
